+++ tests/app/Models/Contact.php
@@ -7,12 +7,13 @@ namespace ShabuShabu\Uid\Tests\App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use ShabuShabu\Uid\Concerns\HasUid;
+use ShabuShabu\Uid\Contracts\Identifiable;
 use ShabuShabu\Uid\Tests\Database\Factories\ContactFactory;
 
 /**
  * @property-read string $uid
  */
-class Contact extends Model
+class Contact extends Model implements Identifiable
 {
     use HasFactory;
     use HasUid;
@@ -22,5 +23,10 @@ class Contact extends Model
     protected static function newFactory(): ContactFactory
     {
         return new ContactFactory;
+    }
+
+    public function uidAlphabet(): string
+    {
+        return 'VbhoG706KLHC2OkEXN5tinPjfwJD3lmupYrRIeMS89ATcUgqszydQ1ZW4aFvBx';
     }
 }